fix: move validation to user/types
